[Feature] AI 서버 API 연결 하기 위한 API 개발

+++ src/main/java/Recommend/Movie/Biorhythm/Dto/BiorhythmAnalysisDTO.java
@@ -8,13 +8,13 @@ import lombok.Getter;
 @Builder
 public class BiorhythmAnalysisDTO {
     @Schema(title = "신체 지수", example = "-51")
-    private int physicalIndex;
+    private double physicalIndex;
 
     @Schema(title = "감성 지수", example = "0")
-    private int emotionalIndex;
+    private double emotionalIndex;
 
     @Schema(title = "지성 지수", example = "97")
-    private int intellectualIndex;
+    private double intellectualIndex;
 
     @Schema(title = "분석 멘트", example = "두뇌 회전이 빠른 날입니다. 몰입감 넘치는 스토리에 도전해보세요.")
     private String statusMessage;
